Add a pkg-config file for libiphb
This allows people to use PKG_CHECK_MODULES() in their configure
scripts.

It's normal to also build (but not install) a foo-uninstalled.pc file,
which allows people to develop against uninstalled versions of the
library. (pkg-config has a special-case for this.) However, that wasn't
possible here: when installed, API users should #include
<iphbd/libiphb.h>; but when uninstalled, the header is in a 'libiphb'
directory. Renaming the directory in the source tree would fix this.