Cache alarm queue state over dsme/device restarts
Alarm queue state is available only as signals sent by timed. If dsme
is restarted for any reason, the alarm state is lost, RTC wakeups do not
get programmed and alarms can be missed due to suspend or power off.

This is fixed by writing alarms state to a file on dsme exit and reading
it back in on dsme startup.

[iphb] Cache alarm queue state over dsme/device restarts