Do not override configured watchdog device paths
When opening the watchdog devices, the code attempted to use on the fly
generated device paths istead of the ones configured in the lookup table.
This could have resulted in situations where error reporting uses device
paths that are not actually used.

As a workaround /dev/watchdog[01] were added to the watchdog list
(these were the paths heuristics could have used) and no error logging
is made from attempts to open non-existing watchdog device nodes.

An error message is still generated if no watchdog device files could
not be opened.