[wlanloader] WLAN loader plugin
Apparently Android by default resets the WLAN driver after
every use of WLAN hotspot. To make it easier to work with android
kernels whose WLAN module depends on this kind of reset service,
this plugin will imitate that behavior by reloading the
kernel module via a systemd service after each use of
WLAN hotspot.

If you want to utilize this plugin, please add a systemd system
service that is named "wlan-module-load.service" to your adaptation
project. The service should modprobe the wlan kernel module at
ExecStart, and rmmod it at ExecStop.

If the systemd service is not present in the system, the plugin will
not do anything.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jolla.com>