Unload plugins in reverse load order
On dsme startup the plugins that provide services are loaded before
the plugins that need those services. On dsme exit, the plugins
are unloaded in the same order as they were loaded - which can
lead to usage of services that are no longer available.

Reverse the module list before unloading them so that the last to
load is the first to unload.